pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prelight"
version = "0.1.0"
description = "Safe data transformation sandbox via MCP — supports Databricks and DuckDB"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "databricks-sql-connector>=3.0.0",
    "duckdb>=0.10.0",
    "sqlglot>=23.0.0",
    "PyGithub>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.scripts]
prelight = "prelight.server:main"

[tool.hatch.build.targets.wheel]
packages = ["prelight"]
